Very basic app. Firebase not working yet.
I decided to remove .gitignore for the Google services since the repo is private, and it will be easier. To switch back to it.

I added you on the firebase project as well with the pggordon email.

Also I can go over this folder structure with you if you want and im open to changes.

Lastly I am defintiely down for suggestions on how to set up this first part, but I have alwasy went the Root widget route that controls the inner workings of the app